fix legacy build after olap, rolap refactoring Signed-off-by: dbulahov <bulahovdenis@gmail.com>
