+++ mondrian/src/test/java/mondrian/rolap/RolapSchemaTest.java
@@ -41,7 +41,7 @@ import org.eclipse.daanse.olap.api.element.Level;
 import org.eclipse.daanse.olap.api.element.Member;
 import org.eclipse.daanse.olap.api.element.OlapElement;
 import org.eclipse.daanse.olap.api.exception.OlapRuntimeException;
-import org.eclipse.daanse.olap.common.RoleImpl;
+import org.eclipse.daanse.olap.access.RoleImpl;
 import org.eclipse.daanse.olap.common.SystemWideProperties;
 import org.eclipse.daanse.olap.exceptions.RoleUnionGrantsException;
 import org.eclipse.daanse.olap.exceptions.UnknownRoleException;
@@ -198,7 +198,7 @@ class RolapCatalogTest {
 
         grant.setCubeGrant(List.of(AccessCubeGrantMappingImpl.builder().build(), AccessCubeGrantMappingImpl.builder().build()));
 
-        org.eclipse.daanse.olap.common.RoleImpl role = new org.eclipse.daanse.olap.common.RoleImpl();
+        org.eclipse.daanse.olap.access.RoleImpl role = new org.eclipse.daanse.olap.access.RoleImpl();
 
         schema.handleCatalogGrant(role, grant);
         assertEquals(org.eclipse.daanse.olap.api.access.AccessCatalog.CUSTOM, role.getAccess(schema));
@@ -217,7 +217,7 @@ class RolapCatalogTest {
         grant.setCube(PhysicalCubeMappingImpl.builder().withName("cube").build());
 
         try {
-            schema.handleCubeGrant(new org.eclipse.daanse.olap.common.RoleImpl(), grant);
+            schema.handleCubeGrant(new org.eclipse.daanse.olap.access.RoleImpl(), grant);
         } catch (OlapRuntimeException e) {
             String message = e.getMessage();
             assertTrue(message.contains(grant.getCube().getName()), message);
@@ -232,7 +232,7 @@ class RolapCatalogTest {
         schema = spy(schema);
         doNothing().when(schema)
             .handleHierarchyGrant(
-                any(org.eclipse.daanse.olap.common.RoleImpl.class),
+                any(org.eclipse.daanse.olap.access.RoleImpl.class),
                 any(RolapCube.class),
                 any(CatalogReader.class),
                 any(AccessHierarchyGrantMappingImpl.class));
@@ -255,7 +255,7 @@ class RolapCatalogTest {
         grant.getDimensionGrants().addAll(List.of(dimensionGrant));
         grant.getHierarchyGrants().addAll(List.of(AccessHierarchyGrantMappingImpl.builder().build()));
 
-        org.eclipse.daanse.olap.common.RoleImpl role = new org.eclipse.daanse.olap.common.RoleImpl();
+        org.eclipse.daanse.olap.access.RoleImpl role = new org.eclipse.daanse.olap.access.RoleImpl();
 
         schema.handleCubeGrant(role, grant);
 
@@ -420,7 +420,7 @@ class RolapCatalogTest {
     {
         RolapCatalog schema = createSchema();
         RolapCube cube = mockCube(schema);
-        org.eclipse.daanse.olap.common.RoleImpl role = new org.eclipse.daanse.olap.common.RoleImpl();
+        org.eclipse.daanse.olap.access.RoleImpl role = new org.eclipse.daanse.olap.access.RoleImpl();
 
         AccessMemberGrantMappingImpl memberGrant = AccessMemberGrantMappingImpl.builder().withMember("member").withAccess(AccessMember.ALL).build();
